feat(release-please.yml): add release-please workflow to automate releases using the release-please-action
The release-please.yml file is added to the .github/workflows directory. This file contains a workflow configuration that automates the release process using the release-please-action. The workflow is triggered on push events to the main branch.

The workflow has the following permissions:
- contents: write
- pull-requests: write

The workflow is named "release-please" and consists of a single job called "release-please". The job runs on the latest version of Ubuntu.

The job has a single step that uses the release-please-action. The action is configured with the following parameters:
- release-type: node
- package-name: release-please-action