chore: improve git workflows for PRs (#35)
* Add lint action

* Change package manager to yarn

* Format files

* chore(lint.yml): update lint workflow to include PR title validation and fix formatting

feat(lint.yml): add PR title validation step using amannn/action-semantic-pull-request@v5 action to enforce semantic PR titles

chore(lint.yml): remove unnecessary permissions configuration for pull requests

chore(lint.yml): fix formatting by adding newline at end of file

* feat(release-please.yml): add release-please workflow to automate releases using the release-please-action

The release-please.yml file is added to the .github/workflows directory. This file contains a workflow configuration that automates the release process using the release-please-action. The workflow is triggered on push events to the main branch.

The workflow has the following permissions:
- contents: write
- pull-requests: write

The workflow is named "release-please" and consists of a single job called "release-please". The job runs on the latest version of Ubuntu.

The job has a single step that uses the release-please-action. The action is configured with the following parameters:
- release-type: node
- package-name: release-please-action

* chore(lint.yml): remove unnecessary step for PR title validation
feat(semantic-pull-request-title.yml): add new workflow file for enforcing semantic pull request titles

* chore(release-please.yml): remove release-please workflow file

The release-please.yml file was deleted as it is no longer needed. This file was responsible for triggering the release-please action on push to the main branch. However, the decision was made to no longer use the release-please action in the project, so the file was removed.